# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-tenant Identity and Access Management",
        "title": "CoCo IAM API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:2026",
    "basePath": "/api/v1",
    "paths": {
        "/admin/organization_users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of users for the given organization (requires organization_id filter) or a single user by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "org-users"
                ],
                "summary": "List or get organization users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID filter (list lookup)",
                        "name": "filter[@exact:organization_id]",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (max 500, default 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.OrgUserListSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a pending org user and sends an activation email. No password on the wire.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "org-users"
                ],
                "summary": "Create organization user",
                "parameters": [
                    {
                        "description": "User payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreateOrgUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.CreateOrgUserSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/admin/organization_users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of users for the given organization (requires organization_id filter) or a single user by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "org-users"
                ],
                "summary": "List or get organization users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (single-user lookup)",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Organization ID filter (list lookup)",
                        "name": "filter[@exact:organization_id]",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (max 500, default 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.OrgUserListSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes the org user from the per-org DB and enqueues ACL cleanup.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "org-users"
                ],
                "summary": "Delete organization user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.OrgUserSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially updates an org user. Username changes are rejected.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "org-users"
                ],
                "summary": "Update organization user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.PatchOrgUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.OrgUserSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.CreateOrgUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "jdoe@example.com"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "organization_id": {
                    "type": "string",
                    "example": "org-uuid"
                },
                "redirect_application_id": {
                    "type": "string",
                    "example": "app-uuid"
                },
                "username": {
                    "type": "string",
                    "example": "jdoe"
                }
            }
        },
        "entity.CreateOrgUserResponse": {
            "type": "object",
            "properties": {
                "activation": {
                    "$ref": "#/definitions/entity.OrgActivationEcho"
                },
                "activation_error": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_a-digi_coco-iam_src_organizations_users_entity.User"
                }
            }
        },
        "entity.CreateOrgUserSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "$ref": "#/definitions/entity.CreateOrgUserResponse"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "entity.OrgActivationEcho": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                }
            }
        },
        "entity.OrgUserListSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.OrgUserResponse"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "entity.OrgUserResponse": {
            "type": "object",
            "properties": {
                "activation_pending": {
                    "type": "boolean",
                    "example": false
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean",
                    "default": true
                },
                "organization_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "entity.OrgUserSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "$ref": "#/definitions/entity.OrgUserResponse"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "entity.PatchOrgUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "new@example.com"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "github_com_a-digi_coco-iam_src_organizations_users_entity.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean",
                    "default": true
                },
                "organization_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "resource not found"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer \u003cJWT token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}